+++ beaconplusWeb/src/components/searchResults/DatasetResultBox.js
@@ -15,7 +15,7 @@ import { svgFetcher } from "../../hooks/fetcher"
 // import BiosamplesStatsDataTable from "./BiosamplesStatsDataTable"
 // import { WithData } from "../Loader"
 import { refseq2chro } from "../Chromosome"
-import { AggregatedPlots } from "../AggregatedPlots"
+import { SummaryPlots } from "../summaries/SummaryPlots"
 
 const HANDOVER_IDS = {
   histoplot: "histoplot",
@@ -241,7 +241,7 @@ export function DatasetResultBox({ data: responseSet, responseMeta, query }) {
       <br/>
       <hr/>
       <h2 className="subtitle has-text-dark">Matched {id} Data  Overview</h2>
-      <AggregatedPlots summaryResults={summaryResults} filterUnknowns={true} />
+      <SummaryPlots summaryResults={summaryResults} filterUnknowns={true} />
       <br/>
       <hr/>
       <h2 className="subtitle has-text-dark">{id} Data Downloads</h2>